+++ core/server/api_container/server/service_network/networking_sidecar/networking_sidecar_wrapper_test.go
@@ -8,8 +8,8 @@ package networking_sidecar
 import (
 	"context"
 	"fmt"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/container_status"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
 	"github.com/stretchr/testify/require"
 	"net"
 	"strings"