+++ core/docs/changelog.md
@@ -521,7 +521,7 @@
 
 # 1.40.1
 ### Fixes
-* Fixed dependency version for `github.com/kurtosis-tech/container-engine-lib` in `launcher/go.sum` and `server/go.sum` files
+* Fixed dependency version for `github.com/kurtosis-tech/kurtosis/container-engine-lib` in `launcher/go.sum` and `server/go.sum` files
 
 # 1.40.0
 ### Breaking Changes